refactor: de-duplicate distribution generators and clear static-analysis findings
Removes duplication and design drift introduced by the recent feature wave
(non-uniform distributions, datafaker realism, check/enum constraints, parallel
fill) and clears actionable JetBrains IntelliJ inspection findings.

DRY / design (behavior-preserving):
- extract the shared CDF binary search into io.bloviate.gen.Cdf; reuse it in
  WeightedCategoricalGenerator and ZipfianIntegerGenerator
- collapse two byte-identical splitmix64 copies (ChildCardinality,
  GroupedPermutationGenerator) onto Mixers.splitmix64
- use Math.clamp for nested min/max (Normal{Double,Integer}Generator,
  DatabaseFiller pool sizing, TPCCConfiguration)
- Comparator.comparing in WeightedCategoricalGenerator

Lifecycle / tests:
- DatabaseFiller.fillParallel uses try-with-resources on the ExecutorService
  for deterministic shutdown
- drop a dead, never-asserted sample list in DistributionGeneratorsTest (the
  test already verifies in-window and recency)

Dead code / nits:
- remove unused TableFiller 4-arg ctor and Builder.table(String); unused logger
  fields; redundant int casts on Math.floorMod; get(0) -> getFirst();
  Optional.orElseGet; assertFalse in three tests
- gitignore the IntelliJ output/ inspection dump

Public API kept intentionally; full reactor build and tests pass.

+++ bloviate-datafaker/src/main/java/io/bloviate/datafaker/Places.java
@@ -68,7 +68,7 @@ public final class Places {
     public static RowContext<Geo> unitedStates(long seed) {
         List<Geo> tuples = UNITED_STATES;
         int size = tuples.size();
-        return new RowContext<>(rowIndex -> tuples.get((int) Math.floorMod(Mixers.splitmix64(seed + rowIndex), size)));
+        return new RowContext<>(rowIndex -> tuples.get(Math.floorMod(Mixers.splitmix64(seed + rowIndex), size)));
     }
 
     /** The loaded reference tuples (package-private; for tests). */
